+++ app/src/main/java/org/oxycblt/auxio/music/Date.kt
@@ -140,8 +140,7 @@ class Date private constructor(private val tokens: List<Int>) : Comparable<Date>
                 min.resolveDate(context)
             }
 
-        override fun equals(other: Any?) =
-            other is Range && min == other.min && max == other.max
+        override fun equals(other: Any?) = other is Range && min == other.min && max == other.max
 
         override fun hashCode() = 31 * max.hashCode() + min.hashCode()